+++ CHANGELOG.md
@@ -6,10 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ##  [0.4.0] - 2024-10-07
-### Agregando
-
 ### Cambios
 - Ajustando codigo en config.yml
+- Cambiando logica para solicitar empate
+- Implementando test para cli.
 
 ##  [0.3.9] - 2024-10-06
 ### Agregando